Completed All of my Tasks
Added a money counter with starting money and updates at the end of a wave (not tested due to inability to kill all goblins) and on kill (tested due to a bug that allowed goblins to hit each other when attacking the base/player. This has been fixed.).

Added a wave mechanic along with UI additions to display the wave progress.

Updated enemy attacks and game over screen.